requirements.txt
argon2-cffi @ file:///C:/ci/argon2-cffi_1613037829118/work
async-generator==1.10
attrs @ file:///tmp/build/80754af9/attrs_1620827162558/work
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
beautifulsoup4==4.9.3
bleach @ file:///tmp/build/80754af9/bleach_1612211392645/work
certifi==2021.5.30
cffi @ file:///C:/ci/cffi_1613247269562/work
click==8.0.1
colorama @ file:///tmp/build/80754af9/colorama_1607707115595/work
cycler==0.10.0
Cython==0.29.23
decorator @ file:///tmp/build/80754af9/decorator_1621259047763/work
defusedxml @ file:///tmp/build/80754af9/defusedxml_1615228127516/work
entrypoints==0.3
imageio==2.9.0
importlib-metadata==4.4.0
ipykernel @ file:///C:/ci/ipykernel_1596206775157/work/dist/ipykernel-5.3.4-py3-none-any.whl
ipython @ file:///C:/ci/ipython_1593446240034/work
ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
ipywidgets @ file:///tmp/build/80754af9/ipywidgets_1610481889018/work
jedi==0.17.0
Jinja2 @ file:///tmp/build/80754af9/jinja2_1621238361758/work
joblib==1.0.1
jsonschema @ file:///tmp/build/80754af9/jsonschema_1602607155483/work
jupyter==1.0.0
jupyter-client @ file:///tmp/build/80754af9/jupyter_client_1616770841739/work
jupyter-console @ file:///tmp/build/80754af9/jupyter_console_1616615302928/work
jupyter-core @ file:///C:/ci/jupyter_core_1612213536848/work
jupyterlab-pygments @ file:///tmp/build/80754af9/jupyterlab_pygments_1601490720602/work
jupyterlab-widgets @ file:///tmp/build/80754af9/jupyterlab_widgets_1609884341231/work
kiwisolver==1.3.1
lxml==4.6.3
MarkupSafe @ file:///C:/ci/markupsafe_1621528313524/work
matplotlib==3.3.4
mistune==0.8.4
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
nbclient @ file:///tmp/build/80754af9/nbclient_1614364831625/work
nbconvert @ file:///C:/ci/nbconvert_1601896775581/work
nbformat @ file:///tmp/build/80754af9/nbformat_1617383369282/work
nest-asyncio @ file:///tmp/build/80754af9/nest-asyncio_1613680548246/work
networkx==2.5.1
nltk==3.6.2
notebook @ file:///C:/ci/notebook_1621528813829/work
numpy==1.19.5
olefile==0.46
opencv-python==4.5.1.48
packaging @ file:///tmp/build/80754af9/packaging_1611952188834/work
pandas==1.1.5
pandocfilters @ file:///C:/ci/pandocfilters_1605120535071/work
parso @ file:///tmp/build/80754af9/parso_1617223946239/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow @ file:///C:/ci/pillow_1617386325740/work
prometheus-client @ file:///tmp/build/80754af9/prometheus_client_1618088486455/work
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1616415428029/work
protobuf==3.15.8
pthflops==0.4.1
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
Pygments @ file:///tmp/build/80754af9/pygments_1621606182707/work
pyparsing @ file:///home/linux1/recipes/ci/pyparsing_1610983426697/work
pyrsistent @ file:///C:/ci/pyrsistent_1600141799440/work
python-crfsuite==0.9.7
python-dateutil @ file:///home/ktietz/src/ci/python-dateutil_1611928101742/work
pytz==2021.1
pyvi==0.1
PyWavelets==1.1.1
pywin32==227
pywinpty==0.5.7
PyYAML==5.4.1
pyzmq==20.0.0
qtconsole @ file:///tmp/build/80754af9/qtconsole_1616775094278/work
QtPy==1.9.0
regex==2021.4.4
scikit-image==0.17.2
scikit-learn==0.24.2
scipy==1.5.4
Send2Trash @ file:///tmp/build/80754af9/send2trash_1607525499227/work
six @ file:///C:/ci/six_1605187303045/work
sklearn-crfsuite==0.3.6
soupsieve==2.2.1
tabulate==0.8.9
tensorboardX==2.2
terminado==0.9.4
testpath @ file:///home/ktietz/src/ci/testpath_1611930608132/work
thop==0.0.31.post2005241907
threadpoolctl==2.1.0
tifffile==2020.9.3
torch==1.2.0
torchvision==0.4.0
tornado @ file:///C:/ci/tornado_1606942379977/work
tqdm==4.61.0
traitlets==4.3.3
typing-extensions==3.10.0.0
wcwidth @ file:///tmp/build/80754af9/wcwidth_1593447189090/work
webencodings==0.5.1
widgetsnbextension==3.5.1
wincertstore==0.2
zipp @ file:///tmp/build/80754af9/zipp_1615904174917/work
